+++ Circles/Views/People/PeopleOverviewScreen.swift
@@ -11,7 +11,7 @@ import Matrix
 
 struct PeopleOverviewScreen: View {
     @ObservedObject var people: ContainerRoom<Matrix.SpaceRoom>
-    @ObservedObject var profile: ContainerRoom<Matrix.Room>
+    @ObservedObject var profile: ProfileSpace
     @ObservedObject var circles: ContainerRoom<CircleSpace>
     @ObservedObject var groups: ContainerRoom<GroupRoom>
     
@@ -108,7 +108,7 @@ struct PeopleOverviewScreen: View {
                 let user = people.session.getUser(userId: room.creator)
                 
                 VStack(alignment: .leading) {
-                    NavigationLink(destination: ConnectedPersonDetailView(space: room),
+                    NavigationLink(destination: ConnectedPersonDetailView(space: room, profile: profile),
                                    tag: user.userId,
                                    selection: $selectedUserId
                     ) {
@@ -248,7 +248,7 @@ struct PeopleOverviewScreen: View {
                     if userId != profile.session.creds.userId {
                         let user = profile.session.getUser(userId: userId)
                         if let friendsSpace = people.rooms.values.first(where: { $0.creator == userId }) {
-                            NavigationLink(destination: ConnectedPersonDetailView(space: friendsSpace),
+                            NavigationLink(destination: ConnectedPersonDetailView(space: friendsSpace, profile: profile),
                                            tag: user.userId,
                                            selection: $selectedUserId
                             ) {