+++ Circles/Views/People/UnconnectedPersonDetailView.swift
@@ -10,8 +10,7 @@ import Matrix
 
 struct UnconnectedPersonDetailView: View {
     @ObservedObject var user: Matrix.User
-    @ObservedObject var myProfileRoom: Matrix.Room
-    @State var mutualFriends: [Matrix.User]? = nil
+    @ObservedObject var myProfileRoom: ProfileSpace
 
     @State private var alertTitle: String = ""
     @State private var alertMessage: String = ""
@@ -66,66 +65,6 @@ struct UnconnectedPersonDetailView: View {
         }
     }
     
-    @ViewBuilder
-    var mutualFriendsSection: some View {
-        LazyVStack(alignment: .leading) {
-            Text("MUTUAL FRIENDS")
-                .font(.subheadline)
-                .foregroundColor(.gray)
-                .padding(.top)
-            if let friends = mutualFriends {
-                if friends.isEmpty {
-                    Text("No mutual friends")
-                        .padding()
-                } else {
-                    ForEach(friends) { friend in
-                        HStack {
-                            UserAvatarView(user: friend)
-                                .frame(width: 80, height: 80)
-                            
-                            VStack(alignment: .leading) {
-                                Text(friend.displayName ?? friend.userId.username)
-                                    .font(.title2)
-                                    .fontWeight(.bold)
-                                Text(friend.userId.stringValue)
-                                    .font(.subheadline)
-                                    .foregroundColor(.gray)
-                            }
-                        }
-                    }
-                }
-            } else {
-                ProgressView()
-            }
-        }
-        .onAppear {
-            Task {
-                let session = user.session
-                // First find the set of circles that we're both in
-                let rooms: [Matrix.Room] = session.rooms.values.filter { room in
-                    room.type == ROOM_TYPE_CIRCLE && room.joinedMembers.contains(user.userId)
-                }
-                // Find the users in those circles who are not (1) me or (2) them
-                let users: Set<Matrix.User> = rooms.reduce([], { curr, room in
-                    let roomMembers: [Matrix.User] = room.joinedMembers.filter {
-                        $0 != user.userId && $0 != user.session.creds.userId
-                    }.compactMap {
-                        session.getUser(userId: $0)
-                    }
-                    return curr.union(roomMembers)
-                })
-                // Sort the set in order to get an Array
-                let sortedUsers = users.sorted(by: {u0,u1 in
-                    u0.userId.stringValue < u1.userId.stringValue
-                })
-                // Update the UI state on the main thread
-                await MainActor.run {
-                    self.mutualFriends = sortedUsers
-                }
-            }
-        }
-    }
-    
     var body: some View {
         ScrollView {
             VStack(alignment: .leading) {
@@ -136,7 +75,7 @@ struct UnconnectedPersonDetailView: View {
                 
                 Divider()
             
-                mutualFriendsSection
+                MutualFriendsSection(user: user, profile: myProfileRoom)
             
             }
         }